[audio] Add 64 bits devfs entries support to audio-driver-ctl
Bug: 502261895
Test: Use the tool in Sherlock

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1567738
Commit-Queue: Andres Oportus <andresoportus@google.com>

+++ src/media/audio/drivers/lib/audio-utils/audio-input.cc
@@ -24,7 +24,7 @@ static constexpr zx::duration kDesiredWakeupPeriod = zx::msec(50);
 static constexpr float MIN_DURATION = 0.100f;
 static constexpr float MAX_DURATION = 86400.0f;
 
-std::unique_ptr<AudioInput> AudioInput::Create(uint32_t dev_id) {
+std::unique_ptr<AudioInput> AudioInput::Create(uint64_t dev_id) {
   fbl::AllocChecker ac;
   std::unique_ptr<AudioInput> res(new (&ac) AudioInput(dev_id));
   if (!ac.check())
